--- packages/functional-tests/src/tests/gltf-animation-test.ts
@@ -1,64 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import * as MRE from '@microsoft/mixed-reality-extension-sdk';
-
-import { Test } from '../test';
-
-export default class GltfAnimationTest extends Test {
-	public expectedResultDescription = "Click to toggle walk animation";
-
-	private assets: MRE.AssetContainer;
-	private animating = false;
-	private prefab: MRE.Actor;
-
-	public cleanup() {
-		this.assets.unload();
-	}
-
-	public async run(root: MRE.Actor): Promise<boolean> {
-		this.assets = new MRE.AssetContainer(this.app.context);
-
-		this.prefab = MRE.Actor.CreateFromGltf(this.assets, {
-			uri: 'https://raw.githubusercontent.com/KhronosGroup/' +
-				'glTF-Sample-Models/master/2.0/CesiumMan/glTF-Binary/CesiumMan.glb',
-			colliderType: 'box',
-			actor: {
-				parentId: root.id,
-				transform: {
-					local: {
-						position: { z: -1 }
-					}
-				}
-			}
-		});
-
-		/* this.prefab.setBehavior(MRE.ButtonBehavior)
-			.onClick(() => this.setAnimationState(this.animating = !this.animating));
-		*/
-
-		await this.prefab.created();
-		for (const anim of this.prefab.animations.values()) {
-			console.log(`animation found: "${anim.name}" (${anim.id})`);
-		}
-
-		this.prefab.setBehavior(MRE.ButtonBehavior)
-			.onClick(() => {
-				const anim = this.prefab.animationsByName.get("animation:0");
-				anim.isPlaying ? anim.stop() : anim.play();
-			});
-
-		await this.stoppedAsync();
-		return true;
-	}
-
-	private setAnimationState(play: boolean): void {
-		if (play) {
-			this.prefab.enableAnimation('animation:0');
-		} else {
-			this.prefab.disableAnimation('animation:0');
-		}
-	}
-}
